+++ Sources/StreamVideo/WebRTC/v2/WebRTCStateAdapter.swift
@@ -43,7 +43,7 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate {
     let peerConnectionFactory: PeerConnectionFactory
     let videoCaptureSessionProvider: VideoCaptureSessionProvider
     let screenShareSessionProvider: ScreenShareSessionProvider
-    let audioSession: StreamAudioSession
+    let audioSession: CallAudioSession
     let trackStorage: WebRTCTrackStorage = .init()
 
     /// Published properties that represent different parts of the WebRTC state.
@@ -124,15 +124,7 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate {
         self.rtcPeerConnectionCoordinatorFactory = rtcPeerConnectionCoordinatorFactory
         self.videoCaptureSessionProvider = videoCaptureSessionProvider
         self.screenShareSessionProvider = screenShareSessionProvider
-        self.audioSession = .init(audioDeviceModule: peerConnectionFactory.audioDeviceModule)
-
-        Task {
-            await configureAudioSession()
-        }
-    }
-
-    deinit {
-        audioSession.dismantle()
+        self.audioSession = .init()
     }
 
     /// Sets the session ID.
@@ -185,7 +177,6 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate {
     /// Sets the WebRTC stats reporter.
     func set(statsAdapter value: WebRTCStatsAdapting?) {
         self.statsAdapter = value
-        value?.audioSession = audioSession
         value?.consume(queuedTraces)
     }
 
@@ -222,6 +213,14 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate {
     func set(isTracingEnabled value: Bool) {
         self.isTracingEnabled = value
         statsAdapter?.isTracingEnabled = value
+    }
+
+    func set(audioSessionPolicy: AudioSessionPolicy) {
+        audioSession.didUpdatePolicy(
+            audioSessionPolicy,
+            callSettings: callSettings,
+            ownCapabilities: ownCapabilities
+        )
     }
 
     // MARK: - Client Capabilities
@@ -363,6 +362,7 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate {
         set(anonymousCount: 0)
         set(participantPins: [])
         trackStorage.removeAll()
+        audioSession.deactivate()
     }
 
     /// Cleans up the session for reconnection, clearing adapters and tracks.
@@ -389,6 +389,7 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate {
         /// We set the initialCallSettings to the last activated CallSettings, in order to maintain the state
         /// during reconnects.
         initialCallSettings = callSettings
+        audioSession.deactivate()
     }
 
     /// Restores screen sharing if an active session exists.
@@ -634,36 +635,21 @@ actor WebRTCStateAdapter: ObservableObject, StreamAudioSessionAdapterDelegate {
         }
     }
 
-    private func configureAudioSession() {
-        audioSession.delegate = self
-
-        $callSettings
-            .removeDuplicates()
-            .sinkTask(storeIn: disposableBag) { [weak audioSession] in
-                do {
-                    try await audioSession?.didUpdateCallSettings($0)
-                } catch {
-                    log.error(error)
-                }
-            }
-            .store(in: disposableBag)
-
-        $ownCapabilities
-            .removeDuplicates()
-            .sinkTask(storeIn: disposableBag) { [weak audioSession] in
-                do {
-                    try await audioSession?.didUpdateOwnCapabilities($0)
-                } catch {
-                    log.error(error)
-                }
-            }
-            .store(in: disposableBag)
+    func configureAudioSession(source: JoinSource?) async throws {
+        audioSession.activate(
+            callSettingsPublisher: $callSettings.removeDuplicates().eraseToAnyPublisher(),
+            ownCapabilitiesPublisher: $ownCapabilities.removeDuplicates().eraseToAnyPublisher(),
+            delegate: self,
+            statsAdapter: statsAdapter,
+            /// If we are joining from CallKit the AudioSession will be activated from it and we
+            /// shouldn't attempt another activation.
+            shouldSetActive: source != .callKit
+        )
     }
 
     // MARK: - AudioSessionDelegate
 
     nonisolated func audioSessionAdapterDidUpdateCallSettings(
-        _ adapter: StreamAudioSession,
         callSettings: CallSettings
     ) {
         Task(disposableBag: disposableBag) { [weak self] in