+++ CHANGELOG.md
@@ -4,8 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+
 ### ✅ Added
 - The SDK now handles the interruptions produced from AVCaptureSession to ensure video capturing is active when needed. [#907](https://github.com/GetStream/stream-video-swift/pull/907)
+
+### 🐞 Fixed
+- AudioSession management issues that were causing audio not being recorded during calls. [#906](https://github.com/GetStream/stream-video-swift/pull/906)
 
 # [1.29.1](https://github.com/GetStream/stream-video-swift/releases/tag/1.29.1)
 _July 25, 2025_